requirements.txt
openai-whisper
torch
Pillow
pytesseract
watchdog
pydub
mutagen
